Moved preparation serializer to apropriate spot.
